add: adds first solution
this is a first draft for a solution to the challenge 0 for the 'advent of vue' challenge. This solution differs a bit from the requirements, as I chose to make the field size and thus the winning combinations dynamic instead of only opting for a 3x3 field and hard coded winning combinations. However, there is still a bug I need to fix (at some point the checks are not started anymore).